+++ NEWS.md
@@ -5,7 +5,7 @@
   [Rogue#33](https://github.com/ms609/Rogue/issues/33).
 - Require R 3.6
 - Remove R.cache dependency: `UnrootedKeys()` now uses a native cache implementation.
-- Remove deprecated function `TreeSplits()`.
+- Remove deprecated functions `TreeSplits()`, `ForestSplits()`, `SplitNumber()`.
 
 
 # TreeTools 1.16.1 (2025-08-24) #